requirements.txt
PyYAML==6.0.1
requests==2.31.0
pytest==7.4.3
